sensorhandler: fix type of get_reservation_id()
The implementation of get_reservation_id() is combining two uint8_t, but
currently returns a uint8_t. Return instead a uint16_t to correct this.

+++ sensorhandler.hpp
@@ -155,7 +155,7 @@ struct GetSdrReq
 namespace request
 {
 
-inline uint8_t get_reservation_id(GetSdrReq* req)
+inline uint16_t get_reservation_id(GetSdrReq* req)
 {
     return (req->reservation_id_lsb + (req->reservation_id_msb << 8));
 };
